PLANET-3711 Fix css issue that prevented user from selecting a checkbox. Fixed js bug in ENForm add/edit pages. Fixed backend bug that prevented ens api call from getting authorized and eventually en form from getting submitted. Removed old ENForm block.
