+++ classes/controller/blocks/class-enblock-controller.php
@@ -49,6 +49,8 @@ if ( ! class_exists( 'ENBlock_Controller' ) ) {
 			add_action( 'admin_print_footer_scripts-post.php', [ $this, 'print_admin_footer_scripts' ], 1 );
 			add_action( 'admin_print_footer_scripts-post-new.php', [ $this, 'print_admin_footer_scripts' ], 1 );
 			add_action( 'admin_enqueue_scripts', [ $this, 'load_admin_assets' ] );
+			add_action( 'wp_ajax_get_en_session_token', [ $this, 'get_session_token' ] );
+			add_action( 'wp_ajax_nopriv_get_en_session_token', [ $this, 'get_session_token' ] );
 		}
 
 		/**
@@ -383,6 +385,32 @@ if ( ! class_exists( 'ENBlock_Controller' ) ) {
 			);
 
 			return $data;
+		}
+
+		/**
+		 * Get en session token for frontend api calls.
+		 */
+		public function get_session_token() {
+			// If this is an ajax call.
+			if ( wp_doing_ajax() ) {
+
+				$nonce    = filter_input( INPUT_POST, '_wpnonce', FILTER_SANITIZE_STRING );
+				$response = [];
+
+				// CSRF protection.
+				if ( ! wp_verify_nonce( $nonce, 'enform_submit' ) ) {
+					$response['message'] = __( 'Invalid nonce!', 'planet4-engagingnetworks' );
+					$response['token']   = '';
+				} else {
+
+					$main_settings     = get_option( 'p4en_main_settings' );
+					$ens_private_token = $main_settings['p4en_frontend_private_api'];
+					$this->ens_api     = new Ensapi( $ens_private_token, false );
+					$token             = $this->ens_api->get_public_session_token();
+					$response['token'] = $token;
+				}
+				wp_send_json( $response );
+			}
 		}
 
 		/**
