--- includes/enform.twig
@@ -1,150 +1,0 @@
-{% block enform %}
-	<section class="block enform-wrap enform-{{ fields.en_form_style }} {% if fields.en_form_style == 'full-width-bg' %}block-footer block-wide{% endif %}">
-		{% if fields.en_form_style == 'full-width-bg' %}
-			<picture>
-				<img src="{{ fields.background_src.0|default(fields.default_image) }}"
-					style="object-position:{{ fields.focus_image }};opacity:{{ fields.opacity }};"
-					border="0"
-					srcset="{{ fields.background_srcset }}"
-					sizes="{{ fields.background_sizes }}"
-				/>
-			</picture>
-		{% endif %}
-
-		<div class="container">
-			<div class="row">
-				<div class="col-md-12">
-					<div class="enform" id="enform" data-redirect-url="{{ redirect_url|default('') }}">
-						{% if fields.title %}
-							<header>
-								<h2 class="page-section-header">{{ fields.title }}</h2>
-							</header>
-						{% endif %}
-						{% if fields.description %}
-							<p class="page-section-description">{{ fields.description|e('wp_kses_post')|raw }}</p>
-						{% endif %}
-						<form id="p4en_form" name="p4en_form" method="post" novalidate>
-							{{ fn( 'wp_nonce_field', nonce_action, '_wpnonce', true, false )|raw }}
-							<input type="hidden" name="enform_submit" value="{{ enform_submit }}">
-							<input type="hidden" name="en_page_id" value="{{ fields.en_page_id }}">
-							<input type="hidden" name="thankyou_title" value="{{ fields.thankyou_title }}">
-							<input type="hidden" name="thankyou_subtitle" value="{{ fields.thankyou_subtitle }}">
-
-							{{ error_msg }}
-							<div class="{% if fields.en_form_style == 'full-width-bg' %}row{% endif %}">
-								<div class="{% if fields.en_form_style == 'full-width-bg' %}col-md-8{% endif %}">
-									<div class="formblock-flex donations-formsection-info">
-										{% for key,data in fields %}
-											{% set errorMessage = __( 'This field is required', 'planet4-engagingnetworks' ) %}
-
-											{% if supporter %}
-												{% if 'title' in data.name %}
-													{% set value = supporter['Title'] %}
-												{% elseif 'first' in data.name %}
-													{% set value = supporter['First name'] %}
-												{% elseif 'last' in data.name %}
-													{% set value = supporter['Last name'] %}
-												{% elseif 'address1' in data.name %}
-													{% set value = supporter['Address 1'] %}
-												{% elseif 'address2' in data.name %}
-													{% set value = supporter['Address 2'] %}
-												{% elseif 'city' in data.name %}
-													{% set value = supporter.City %}
-												{% elseif 'country' in data.name %}
-													{% set value = supporter.Country %}
-												{% elseif 'post' in data.name %}
-													{% set value = supporter.Postcode %}
-												{% elseif 'email' in data.name %}
-													{% set value = supporter.Email %}
-												{% elseif 'region' in data.name %}
-													{% set value = supporter.County %}
-												{% elseif 'phoneNumber' in data.name %}
-													{% set value = supporter['Phone Number'] %}
-												{% elseif 'Birth' in data.name %}
-													{% set value = supporter['birthday'] %}
-												{% else %}
-													{% set value = '' %}
-												{% endif %}
-											{% else %}
-												{% set value = '' %}
-											{% endif %}
-
-											{% if data.type == 'email' %}
-												{% set errorMessage = __( 'Please enter a valid e-mail address.', 'planet4-engagingnetworks' ) %}
-											{% elseif data.type == 'country' %}
-												{% set errorMessage = __( 'Please select a country.', 'planet4-engagingnetworks' ) %}
-											{% endif %}
-
-											{% if 'en_page_id' != key and 'en_form_style' != key and '' != data.value %}
-												{% if 'text' == data.type or 'email' == data.type %}
-													<div class="en__field en__field--text en__field--{{ data.id }} en__field--{{ data.name }}">
-														<div class="en__field__element en__field__element--text form-group" style="display: block;">
-															<input id="en__field_supporter_{{ data.name }}"
-																name="supporter.{{ data.name }}"
-																type="{{ 'phoneNumber' == data.name ? 'tel' : data.type }}"
-																class="en__field__input en__field__input--text form-control"
-																value="{{ value }}"
-																data-errormessage="{{ errorMessage }}"
-																placeholder="{{ data.label }}{{ 'Birth' in data.name ? ' (yyyy-mm-dd)' : '' }}{{ 'true' == data.mandatory ? '*' : '' }}"
-																{{ 'true' == data.mandatory or 'email' == data.type ? 'required' : '' }}
-																size="40" />
-														</div>
-													</div>
-												{% elseif 'country' == data.type %}
-													{% include 'country_select.twig' with { errorMessage: errorMessage } %}
-												{% endif %}
-											{% endif %}
-										{% endfor %}
-									</div>
-									{% if supporter %}
-										{% for key,data in supporter.questions %}
-											{% if 'GEN' == data.type %}
-												<div style="display: flex; flex-direction: row;">
-													<label for="en__field_supporter_questions_{{ data.id }}"
-														   class="en__field__label en__field__label--item form-check-label"
-														   style="color: white; margin-right: 20px;">{{ data.label }}{{ 'true' == data.mandatory ? '*' : '' }}</label>
-													<input id="en__field_supporter_questions_{{ data.id }}"
-														   name="supporter.questions.{{ data.id }}"
-														   type="text"
-														   class="en__field__input en__field__input--text"
-														   style="flex: 1;"
-														   value=""
-															{{ 'true' == data.mandatory ? 'required' : '' }} />
-												</div>
-											{% endif %}
-										{% endfor %}
-										{% for key,data in supporter.questions %}
-											{% if 'OPT' == data.type %}
-												<div class="form-check-label-block custom-control">
-													<input id="en__field_supporter_questions_{{ data.id }}"
-														   name="supporter.questions.{{ data.id }}"
-														   class="en__field__input en__field__input--checkbox"
-														   type="checkbox"
-														   value="{{ data.value }}" {{ 'Y' == data.value ? 'checked' : '' }}
-														   {{ 'true' == data.mandatory ? ' required' : '' }} />
-													<label for="en__field_supporter_questions_{{ data.id }}"
-														   class="custom-control-description form-check-label">
-														   {{ data.label }}{{ 'true' == data.mandatory ? '*' : '' }}
-													</label>
-												</div>
-											{% endif %}
-										{% endfor %}
-									{% endif %}
-									<p>* {{ __( 'Required', 'planet4-engagingnetworks' ) }}</p>
-								</div>
-
-								<div class="{% if fields.en_form_style == 'full-width-bg' %}col-md-4{% endif %} submit">
-									<button type="submit" form="p4en_form" name="p4en_form_save_button" id="p4en_form_save_button" class="btn btn-primary btn-block {% if fields.en_form_style != 'full-width-bg' %}w-auto{% endif %}" >{{ '' != fields.button_text ? fields.button_text : __( 'Sign', 'planet4-engagingnetworks' ) }}
-										<div class="three-quarters en-spinner" id="spinner" style="display:none;">...</div>
-									</button>
-									<div class="enform-notice"></div>
-								</div>
-							</div>
-						</form>
-					</div>
-					<div class="clear"></div>
-				</div>
-			</div>
-		</div>
-	</section>
-{% endblock %}
