+++ includes/enform_post.twig
@@ -68,27 +68,33 @@
 						</div>
 					</div>
 				{% elseif 'checkbox' == field.input_type %}
-					<div class="en__field en__field--check en__field--{{ field.id }}">
-						<div class="en__field__element en__field__element--check form-group form-check-label-block custom-control p4-custom-control-input">
-							<label class="custom-checkbox">
-								{% if 'GEN' == field.en_type %}
-									<span class="custom-control-description">
+					{% if 'GEN' == field.en_type %}
+						<span class="custom-control-description">
 										{{ field.label|e('wp_kses_post')|raw }}
 									</span><br />
-									{% for value, question_option in field.question_options|cast_to_array %}
+						{% for value, question_option in field.question_options|object_to_array %}
+							<div class="en__field en__field--check en__field--{{ field.id }}">
+								<div class="en__field__element en__field__element--check form-group form-check-label-block custom-control p4-custom-control-input">
+									<label class="custom-checkbox">
 										<input
-												id="en__field_supporter_{{ field.name }}"
-												name="{{ en_input_name }}"
-												type="checkbox"
-												class="en__field__input en__field__input--checkbox"
-												value="{{ value }}"
-												{{ question_option.selected ?  ' checked' : '' }}
+											id="en__field_supporter_{{ field.name }}"
+											name="{{ en_input_name }}"
+											type="checkbox"
+											class="en__field__input en__field__input--checkbox"
+											value="{{ value }}"
+											{{ question_option.selected ?  ' checked' : '' }}
 										/>
 										<span class="custom-control-description">
 											{{ question_option.label|e('wp_kses_post')|raw }}
-										</span>
-									{% endfor %}
-								{% elseif 'OPT' == field.en_type %}
+										</span><br />
+									</label>
+								</div>
+							</div>
+						{% endfor %}
+					{% elseif 'OPT' == field.en_type %}
+						<div class="en__field en__field--check en__field--{{ field.id }}">
+							<div class="en__field__element en__field__element--check form-group form-check-label-block custom-control p4-custom-control-input">
+								<label class="custom-checkbox">
 									<input
 										id="en__field_supporter_{{ field.name }}"
 										name="{{ en_input_name }}"
@@ -101,10 +107,10 @@
 									<span class="custom-control-description">
 										{{ field.label|e('wp_kses_post')|raw }}{{ 'true' == field.required ? '*' : '' }}
 									</span>
-								{% endif %}
-							</label>
+								</label>
+							</div>
 						</div>
-					</div>
+					{% endif %}
 				{% elseif 'country' == field.input_type %}
 					<div class="en__field en__field--select en__field--{{ field.id }}">
 						<div class="en__field__element en__field__element--select form-group">
